% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascen.R
\name{datascen}
\alias{datascen}
\title{datascen}
\usage{
datascen(
  G_2p6_d,
  G_6p0_d,
  G_8p5_d,
  I_2p6_d,
  I_6p0_d,
  I_8p5_d,
  H_2p6_d,
  H_6p0_d,
  H_8p5_d,
  M_2p6_d,
  M_6p0_d,
  M_8p5_d,
  N_2p6_d,
  N_6p0_d,
  N_8p5_d,
  G_hist_d,
  I_hist_d,
  H_hist_d,
  N_hist_d,
  M_hist_d
)
}
\arguments{
\item{M_hist_d}{}
}
\value{
Returns a list in a tidy order. Each entry of the list describes one river and combines the date, the model and the scenario of the projection in a dataframe.
}
\description{
Function combines all projections in one tidy dataset (list). The function was written for the approach to combine different models in the analysis to reduce or understand uncertainities.
}
\examples{
\dontrun{mod_data=datascen(G_2p6_d, G_6p0_d, G_8p5_d, I_2p6_d, I_6p0_d, I_8p5_d, H_2p6_d, H_6p0_d, H_8p5_d, M_2p6_d, M_6p0_d, M_8p5_d, N_2p6_d, N_6p0_d, N_8p5_d, G_hist_d, I_hist_d, H_hist_d, N_hist_d, M_hist_d)}
}
