% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdc_reader.R
\name{grdc_reader}
\alias{grdc_reader}
\title{GRDC-Data List}
\usage{
grdc_reader(metadata, path)
}
\arguments{
\item{metadata}{Data Frame; Overview of GRDC-Dataset.  The metadata can be created by \link[lfanalyse]{metadata_grdc} function.}

\item{path}{character; Pathway to local grdc data folder.}
}
\value{
Grdc List. Each entry contains discharge measurements as well as the corresponding date of a specific station. The number of the displayed stations depends on the rownumber of the metadataset.
}
\description{
Function returns list. It may take some time to load the GRDC-List, depending on the amount of stations within the metadataset.
}
\examples{
\dontrun{
grdc_list(metadata_germany, "/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/")
}


}
