% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfanalyse.R
\docType{package}
\name{lfanalyse}
\alias{lfanalyse}
\title{Low Flow Analysis Package}
\description{
R-Package to analyse discharge data. To get startet "lfanalyse" offers fucntions to load and structure the GRDC-Dataset. For the analysis package contains functions for a descriptive overview, a threshold based analysis as well as a low flow trend analysis.
}
\author{
Mai-Britt Berghöfer \ermail{berghoefer@uni-potsdam.de}
}
