% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_repg.R
\name{metadata_repg}
\alias{metadata_repg}
\title{Representative Stations in Germany}
\usage{
metadata_repg(metadata, mark = F)
}
\arguments{
\item{metadata}{"matrix" "array" ;  Metadata of grdc_dataset. Can be created by \link[lfanalyse]{metadata_grdc} function.}

\item{mark}{logical; default=F; if FALSE: Function returns new, metadata data.frame containing representative stations only. If TRUE: function returns same metadata data.frame (of GRDC- Dataset) and adds a new column that identifies whether a station is representative (T) or not (F).}
}
\value{
"matrix". Same arrangement like metadata. Contains representative stations or marks representative stations in original/input metadata.
}
\description{
Function is searching in input-metadata for representative stations. Depending on the settings, if mark=T it returns the original metadata and adds a column containing logical values. If one entry in the column is TRUE- the station is representative, FALSE if its not. Or, if mark=F it returns a metadata data.frame only containing the representative stations.
}
\examples{
\dontrun{metadata_repg(metadata_germany)
metadata_germany=metadata_repg(metadata_germany, mark=T)}


}
