% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qplot.R
\name{Qplot}
\alias{Qplot}
\title{Discharge Plot of the entire Time Series}
\usage{
Qplot(data, station, pettitt = F)
}
\arguments{
\item{data}{list; Contains all stations that the discharge analysis should consider. List can be created by \link[lfanalyse]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{pettitt}{logical; (default=F); Pettitt-Test. If TRUE : non-parametric test applied on discharge measurements. Tests for a "shift in the central tendency of a time series" (source:\link[trend]{pettitt.test} ).}
}
\value{
Graphic ; Function visualizes discharge time series.
}
\description{
Desciptive statistics. Time series of discharge at specific station. Including all discharge measurements.
}
\examples{
\dontrun{ Qplot(mosel, "COCHEM", T)}


}
