% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raster.R
\name{Rasterplot}
\alias{Rasterplot}
\title{Rasterplot of a Discharge Time Series}
\usage{
Rasterplot(station, data)
}
\arguments{
\item{station}{character; Name of the station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; Contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}
}
\value{
Graphic ; Function visualizes discharge time series as a rasterplot.
}
\description{
Function displays discharge Values at a station. Rasterplot covers entire time series
}
\examples{
\dontrun{ Rasterplot("COCHEM", data )}
}
