% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalproj.R
\name{seasonalproj}
\alias{seasonalproj}
\title{Monthly resolution of discharge value projections}
\usage{
seasonalproj(rivernames, sel_names, mod_data, NF, FF, scenario)
}
\arguments{
\item{rivernames}{vector; contains rivernames as they are listed in the dataset (mod_data)}

\item{sel_names}{vector; contains rivernames as they appear in the plot}

\item{mod_data}{list in a tidy order, as it is generated by \link[lfanalyse]{datascen} function. List combines all rivers and projected values.}

\item{NF}{numerical; vector containing all the years that account for the near future. The function itself operates with hydrological years.}

\item{FF}{numerical; vector containing all the years that account for the far future. The function itself operates with hydrological years.}

\item{scenario}{charcter, One out of 3 representative RCP Scnearios (2p6, 6p0, 8p5).}

\item{P}{numerical; vector containing all the years that account for the past. The function itself operates with hydrological years.}
}
\value{
seasonal plot with monthly resolutio
}
\description{
Function allows to compare seasonality of monthly averaged minimum dischargeva lues for 3 rivers and one scenario.
}
\examples{
\dontrun{pl=seasonalproj(rivernames, sel_names , mod_data)}
}
